requirements.txt
faster-whisper>=1.0.3
sounddevice>=0.4.6
numpy>=1.26,<2.0
pynput>=1.7.7
pystray>=0.19.5
Pillow>=10.0
pyperclip>=1.8.2
nvidia-cublas-cu12
nvidia-cudnn-cu12==9.*
